+++ app/api/cron/expenses/route.ts
@@ -1,6 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
+import { toErrorResponse } from "@/lib/errors";
 import { listAllUsers, adminListExpenses, type AdminUser } from "@/lib/firebase-admin";
-import { getIstDateString, hasCronBeenSentToday, markCronAsSentToday } from "@/lib/cron-guard";
+import { hasCronBeenSentToday, markCronAsSentToday } from "@/lib/cron-guard";
+import { getIstDateString } from "@/lib/dates";
 
 export const dynamic = "force-dynamic";
 
@@ -250,7 +252,6 @@ export async function POST(req: NextRequest) {
     const sentCount = results.filter((r) => r.sent).length;
     return NextResponse.json({ success: true, period, usersProcessed: users.length, emailsSent: sentCount, results });
   } catch (error: any) {
-    console.error("Error in cron/expenses:", error);
-    return NextResponse.json({ error: error.message || "Failed to run expense summary cron" }, { status: 500 });
+    return toErrorResponse(error, "Error in cron/expenses");
   }
 }
